More builders and cleanup code Add support for packet notifications

+++ src/main/java/me/moros/bending/ability/air/AirBlade.java
@@ -147,7 +147,7 @@ public class AirBlade extends AbilityInstance {
     if (deltaTime >= userConfig.maxChargeTime) {
       factor = userConfig.chargeFactor;
     } else if (deltaTime > 0.3 * userConfig.maxChargeTime) {
-      double deltaFactor = (userConfig.chargeFactor - factor) * deltaTime / (double) userConfig.maxChargeTime;
+      double deltaFactor = (userConfig.chargeFactor - factor) * deltaTime / userConfig.maxChargeTime;
       factor += deltaFactor;
     }
     charging = false;
